requirements.txt
docling==2.30.0
fastapi==0.115.12
fitz==0.0.1.dev2
gradio==5.25.0
langchain==0.3.23
langchain_community==0.3.21
langchain_huggingface==0.1.2
langchain_openai==0.3.12
numpy==2.2.4
openai==1.73.0
opencv_python==4.10.0.84
pandas==2.2.3
pdf2image==1.17.0
pydantic==2.11.3
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.1.0
sentence_transformers==3.2.1
torch==2.5.1+cu121
tqdm==4.66.5
tqdm==4.67.0
transformers==4.46.2
uvicorn==0.34.1
